requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
daphne==4.1.2
Django==5.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
supabase==2.15.0